+++ core/src/gas.rs
@@ -18,6 +18,8 @@
 
 //! Gas module.
 
+use crate::env::Ext;
+
 /// This trait represents a token that can be used for charging `GasCounter`.
 ///
 /// Implementing type is expected to be super lightweight hence `Copy` (`Clone` is added
@@ -290,6 +292,38 @@ impl GasAllowanceCounter {
         let new_value = self.0.checked_add(amount as u128);
 
         self.0 = new_value.unwrap_or(u128::MAX);
+    }
+}
+
+/// Provide functionality to safely make gas refund after charge,
+/// it checks, that it's not refunded more gas than has been charged.
+pub struct GasRefunder {
+    amount: u64,
+}
+
+/// [GasRefunder] error, in case somebody tries to refund more than was charged.
+#[derive(Debug, derive_more::Display)]
+#[display(fmt = "Trying to refund {_1}, but {_0} was charged")]
+pub struct RefundError(u64, u64);
+
+impl GasRefunder {
+    /// Charge gas.
+    pub fn charge<E: Ext>(ext: &mut E, amount: u64) -> Result<Self, E::Error> {
+        ext.charge_gas(amount)?;
+        Ok(Self { amount })
+    }
+
+    /// Refund gas.
+    pub fn refund<E: Ext>(
+        self,
+        ext: &mut E,
+        amount: u64,
+    ) -> Result<Result<(), E::Error>, RefundError> {
+        if self.amount < amount {
+            return Err(RefundError(self.amount, amount));
+        }
+
+        Ok(ext.refund_gas(amount))
     }
 }
